nexus: Implement NVMe Admin Passthru
Add NVME_ADMIN as a permitted IO type and implement NVMe Admin
Passthru starting from BdevHandle when creating a snapshot, which
sends an admin command with opcode c0h, until it reaches nexus_bdev.

Currently this passes it through to the underlying device, which fails
to handle it, so it requires further plumbing to the replica layer.